py-modules/cli/pyproject.toml
[tool.poetry]
authors = ["Daven Quinn <dev@davenquinn.com>"]
description = "A command-line application to control the Mapboard system"
name = "mapboard.cli"
packages = [{ include = "mapboard" }]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
click = "^8.1.3"
"macrostrat.app_frame" = "^1.2.3"
"macrostrat.database" = "^3.3.0"
"macrostrat.dinosaur" = "^3.0.1"
"macrostrat.utils" = "^1.2.1"
"mapboard.topology-manager" = { path = "../topology-manager", develop = true }
pytest = "^7.3.1"
python = "^3.11"
python-daemon = "^3.0.1"
zeroconf = "^0.56.0"
ipython = "^8.14.0"
migra = "^3.0.1647431138"
rich = "^13.4.2"
geopandas = "^0.14.4"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
pytest = "^7.3.0"

[tool.poetry.scripts]
mapboard = "mapboard.cli:app"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
